Fold auto-initialization tests into the HandleSynths test file
The machinery lives in HandleSynths, so its tests do too — the
SupportAutoInitializedProperties folder is gone entirely.
